refactor bags_api
